Weather-Data-ETL-Pipeline/requirements.txt
requests
pandas
python-dateutil
geopy==2.3.0
boto3
python-dotenv